hoai2025: share link behind DCDO flag
