+++ apps/src/lab2/views/dialogs/ShareDialog.tsx
@@ -5,7 +5,7 @@ import Modal from '@code-dot-org/component-library/modal';
 import Typography from '@code-dot-org/component-library/typography';
 import classNames from 'classnames';
 import QRCode from 'qrcode.react';
-import React, {useCallback, useState} from 'react';
+import React, {useCallback} from 'react';
 import FocusLock from 'react-focus-lock';
 
 import {hideShareDialog} from '@cdo/apps/code-studio/components/shareDialogRedux';
@@ -17,11 +17,10 @@ import analyticsReporter from '@cdo/apps/metrics/AnalyticsReporter';
 import {SignInState} from '@cdo/apps/templates/currentUserRedux';
 import {SubmissionStatusType} from '@cdo/apps/templates/projects/submitProjectDialog/submitProjectApi';
 import {commonI18n as i18n} from '@cdo/apps/types/locale';
-import copyToClipboard from '@cdo/apps/util/copyToClipboard';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
-import trackEvent from '@cdo/apps/util/trackEvent';
 import {ProjectSubmissionStatus} from '@cdo/generated-scripts/sharedConstants';
 
+import {CopyToClipboardButton} from './CopyToClipboardButton';
 import HoaiCongrats from './finishDialogs/HoaiCongrats';
 
 import moduleStyles from './share-dialog.module.scss';
@@ -30,44 +29,6 @@ const TEACHER_FEEDBACK_LINK =
   'https://docs.google.com/forms/d/e/1FAIpQLSflGeMmY_ff1QllJfpTsWGZdn_xv6dKpPba_evTMwfbvG3FTA/viewform';
 const STUDENT_FEEDBACK_LINK =
   'https://docs.google.com/forms/d/e/1FAIpQLSeZGNgX4wDvA29stId_Q2toofJN-r12zSP8yBMZ-E9KW5XPWg/viewform';
-
-const CopyToClipboardButton: React.FunctionComponent<{
-  shareUrl: string;
-  projectType: ProjectType;
-  channelId: string | undefined;
-}> = ({shareUrl, projectType, channelId}) => {
-  const [copiedToClipboard, setCopiedToClipboard] = useState(false);
-
-  const handleCopyToClipboard = useCallback(() => {
-    copyToClipboard(shareUrl, () => {
-      setCopiedToClipboard(true);
-    });
-    trackEvent('share', 'share_copy_url', {value: projectType});
-    analyticsReporter.sendEvent(
-      EVENTS.SHARING_LINK_COPY,
-      {
-        lab_type: projectType,
-        channel_id: channelId,
-      },
-      PLATFORMS.STATSIG
-    );
-  }, [shareUrl, projectType, channelId]);
-
-  return (
-    <Button
-      iconLeft={{
-        iconName: copiedToClipboard ? 'clipboard-check' : 'clipboard',
-      }}
-      ariaLabel={i18n.copyLinkToProject()}
-      text={i18n.copyLinkToProject()}
-      type="secondary"
-      color="black"
-      size="m"
-      onClick={handleCopyToClipboard}
-      className={moduleStyles.shareDialogButton}
-    />
-  );
-};
 
 const AfeCareerTourBlock: React.FunctionComponent = () => {
   const careersUrl =
@@ -196,6 +157,9 @@ const ShareDialog: React.FunctionComponent<{
       <HoaiCongrats
         handleClose={handleClose}
         finishUrl={finishUrl}
+        shareUrl={shareUrl}
+        projectType={projectType}
+        channelId={channelId}
         theme={theme}
       />
     );
