+++ lib/dynamic_config/dcdo.rb
@@ -71,7 +71,8 @@ class DCDOBase < DynamicConfigBase
       'actioncable-disconnect-timeout': DCDO.get('actioncable-disconnect-timeout', 30 * 1000),
       'disable-progress-v1': DCDO.get('disable-progress-v1', false),
       'detect-remote-network-config': DCDO.get('detect-remote-network-config', {}),
-      'show-aita-lesson-summaries': DCDO.get('show-aita-lesson-summaries', false)
+      'show-aita-lesson-summaries': DCDO.get('show-aita-lesson-summaries', false),
+      'hoai2025-share-enabled': DCDO.get('hoai2025-share-enabled', false)
     }
   end
 end
